similarity/requirements.txt
numpy==1.22.4
openreview-py==1.2.4
sentence-transformers==2.2.0
stanza==1.4.0
tqdm==4.64.0
